Report Layout Unification with seperate report layout component
